Updated package.json
Interesting read in the Knowledge section. Using "node .", it will look for index.js which doesn't exist. By specifying the exact name it will use that file.

Tested locally using the AWS backend and the API works as intended which leads me to believe that this might be causing some of the issues